refactor(logging): move logger module
