+++ app/pdfreader/pdf_loader.py
@@ -7,7 +7,7 @@ from typing import List, Tuple
 import pdfplumber
 
 from ..tokenizer import Tokenizer
-from ..logger import logger
+from logs.logger import logger
 
 
 def load_pdf(path: str) -> Tuple[str, str]:
