+++ app/tokenizer.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from .logger import logger
+from logs.logger import logger
 
 try:
     import tiktoken
