+++ app/wowrunner.py
@@ -4,7 +4,7 @@ from typing import Optional
 from .openai_batch import run_batch, poll_batch_until_complete
 from utils.formatters import format_batch_summary, format_batch_results, calculate_cost_estimate
 from . import prompt_store
-from .logger import logger
+from logs.logger import logger
 
 
 class WowRunner:
