+++ logs/__init__.py
@@ -1,0 +1,4 @@
+from .logger import logger, setup_logger
+
+__all__ = ["logger", "setup_logger"]
+
